# dennic_api_gateway/api/docs/swagger.yaml
definitions:
...
  model_booking_service.Archive:
    properties:
      created_at:
        type: string
      doctor_availability_id:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      patient_problem:
        type: string
      payment_amount:
        type: number
      payment_type:
        type: string
      start_time:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  model_booking_service.ArchivesType:
    properties:
      archives:
        items:
          $ref: '#/definitions/model_booking_service.Archive'
        type: array
      count:
        type: integer
    type: object
...
  model_booking_service.CreateArchiveReq:
    properties:
      doctor_availability_id:
        type: integer
      end_time:
        type: string
      patient_problem:
        type: string
      payment_amount:
        type: number
      payment_type:
        type: string
      start_time:
        type: string
      status:
        type: string
    type: object
...
  model_booking_service.UpdateArchiveReq:
    properties:
      archive_id:
        type: string
      doctor_availability_id:
        type: integer
      end_time:
        type: string
      patient_problem:
        type: string
      payment_amount:
        type: number
      payment_type:
        type: string
      start_time:
        type: string
      status:
        type: string
    type: object
...
  model_common.ResponseError:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  model_common.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/model_common.ResponseError'
    type: object
...
  models.StatusRes:
    properties:
      status:
        type: boolean
    type: object
host: dennic.uz:9050
info:
  contact: {}
  title: Dennic Project
  version: "1.7"
paths:
...
  /v1/archive:
    delete:
      consumes:
      - application/json
      description: DeleteArchive - Api for delete archive
      parameters:
      - in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StatusRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: DeleteArchive
      tags:
      - Archive
    get:
      consumes:
      - application/json
      description: ListArchive - Api for list archive
      parameters:
      - description: searchField
        enum:
        - status
        in: query
        name: searchField
        type: string
      - in: query
        name: limit
        type: string
      - in: query
        name: order_by
        type: string
      - in: query
        name: page
        type: string
      - in: query
        name: value
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.ArchivesType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: ListArchive
      tags:
      - Archive
    post:
      consumes:
      - application/json
      description: CreateArchive - Api for crete archive
      parameters:
      - description: CreateArchiveReq
        in: body
        name: CreateArchiveReq
        required: true
        schema:
          $ref: '#/definitions/model_booking_service.CreateArchiveReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.Archive'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: CreateArchive
      tags:
      - Archive
    put:
      consumes:
      - application/json
      description: UpdateArchive - Api for update archive
      parameters:
      - description: UpdateArchiveReq
        in: body
        name: UpdateArchiveReq
        required: true
        schema:
          $ref: '#/definitions/model_booking_service.UpdateArchiveReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.Archive'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: UpdateArchive
      tags:
      - Archive
  /v1/archive/get:
    get:
      consumes:
      - application/json
      description: GetArchive - Api for get archive
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.Archive'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: GetArchive
      tags:
      - Archive
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"